Complete revision of controller
