+++ core/config/routing.yml
@@ -1,44 +1,50 @@
 consim_core_index:
     path: /game
-    defaults: { _controller: consim.core.controller.Index:display }
+    defaults: { _controller: consim.core.controller.location:showLocation }
 
 consim_core_location:
     path: /game/location/{location_id}
-    defaults: { _controller: consim.core.controller.Index:showLocation, location_id: 0 }
+    defaults: { _controller: consim.core.controller.location:showLocation, location_id: 0 }
     requirements:
         location_id: \w+
 
 consim_core_building:
     path: /game/location/{location_id}/building/{building_id}
-    defaults: { _controller: consim.core.controller.Index:showLocationBuilding, location_id: 0, building_id: 0 }
+    defaults: { _controller: consim.core.controller.location:showLocationBuilding, location_id: 0, building_id: 0 }
     requirements:
         location_id: \w+
         building_id: \w+
 
 consim_core_activity:
     path: /game/activity
-    defaults: { _controller: consim.core.controller.Index:showActivity}
+    defaults: { _controller: consim.core.controller.action:showActivity}
 
-consim_core_action:
-    path: /game/action/{action_id}
-    defaults: { _controller: consim.core.controller.Index:showAction, action_id: 0}
-    requirements:
-            action_id: \w+
+consim_core_work_end:
+    path: /game/work/end
+    defaults: { _controller: consim.core.controller.request:endWork }
 
-consim_core_travel:
-    path: /game/action/travel/{travel_id}
-    defaults: { _controller: consim.core.controller.Action:travel, travel_id: 0 }
+consim_core_work_start:
+    path: /game/work/start
+    defaults: { _controller: consim.core.controller.request:startWork }
+
+consim_core_travel_start:
+    path: /game/travel/start/{travel_id}
+    defaults: { _controller: consim.core.controller.request:startTravel, travel_id: 0 }
     requirements:
         travel_id: \w+
 
-consim_core_work_end:
-    path: /game/action/work/end
-    defaults: { _controller: consim.core.controller.Action:endWork }
+consim_core_travel:
+    path: /game/travel/{travel_id}
+    defaults: { _controller: consim.core.controller.action:showTravel, travel_id: 0}
+    requirements:
+        travel_id: \w+
 
-consim_core_work_start:
-    path: /game/action/work/start
-    defaults: { _controller: consim.core.controller.Action:startWork }
+consim_core_work:
+    path: /game/work/{work_id}
+    defaults: { _controller: consim.core.controller.action:showWork, work_id: 0}
+    requirements:
+        work_id: \w+
 
 consim_core_register:
     path: /game/register
-    defaults: { _controller: consim.core.controller.Register:display }
+    defaults: { _controller: consim.core.controller.register:display }
