+++ core/config/services/controller.yml
@@ -1,0 +1,65 @@
+imports:
+    - { resource: tables.yml }
+
+services:
+    consim.core.controller.location:
+            class: consim\core\controller\LocationController
+            arguments:
+                - @config
+                - @service_container
+                - @controller.helper
+                - @user
+                - @template
+                - @request
+                - @dbal.conn
+
+    consim.core.controller.action:
+        class: consim\core\controller\ActionController
+        arguments:
+            - @config
+            - @service_container
+            - @controller.helper
+            - @user
+            - @template
+            - @request
+
+    consim.core.controller.request:
+        class: consim\core\controller\RequestController
+        arguments:
+            - @service_container
+            - @controller.helper
+            - @user
+            - @request
+
+    consim.core.controller.register:
+        class: consim\core\controller\RegisterController
+        arguments:
+            - @config
+            - @service_container
+            - @controller.helper
+            - @user
+            - @template
+            - @request
+            - @dbal.conn
+
+    consim.core.controller.acpnews:
+        class: consim\core\controller\ACP_News
+        arguments:
+            - @config
+            - @service_container
+            - @controller.helper
+            - @user
+            - @template
+            - @request
+            - @dbal.conn
+
+    consim.core.controller.acpanniversary:
+        class: consim\core\controller\ACP_Anniversary
+        arguments:
+            - @config
+            - @service_container
+            - @controller.helper
+            - @user
+            - @template
+            - @request
+            - @dbal.conn
