+++ core/config/services/entity.yml
@@ -1,0 +1,155 @@
+imports:
+    - { resource: tables.yml }
+
+services:
+    consim.core.entity.building:
+            class: consim\core\entity\Building
+            scope: prototype # scope MUST be prototype for this to work!
+            arguments:
+                - @dbal.conn
+                - %tables.consim.buildings%
+                - %tables.consim.building_types%
+
+    consim.core.entity.action:
+        class: consim\core\entity\Action
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - @service_container
+            - %tables.consim.actions%
+            - %tables.consim.user%
+            - %tables.consim.works%
+            - %tables.consim.inventory_items%
+
+    consim.core.entity.route:
+        class: consim\core\entity\Route
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.routes%
+
+    consim.core.entity.province:
+        class: consim\core\entity\Province
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.provinces%
+
+    consim.core.entity.route_location:
+        class: consim\core\entity\RouteLocation
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.routes%
+            - %tables.consim.locations%
+            - %tables.consim.location_types%
+            - %tables.consim.provinces%
+            - %tables.consim.countries%
+
+    consim.core.entity.location:
+        class: consim\core\entity\Location
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.locations%
+            - %tables.consim.location_types%
+            - %tables.consim.provinces%
+            - %tables.consim.countries%
+
+    consim.core.entity.location_type:
+        class: consim\core\entity\LocationType
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.location_types%
+
+    consim.core.entity.work:
+        class: consim\core\entity\Work
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - @service_container
+            - %tables.consim.works%
+            - %tables.consim.skills%
+
+    consim.core.entity.work_output:
+        class: consim\core\entity\WorkOutput
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.work_outputs%
+            - %tables.consim.items%
+
+    consim.core.entity.news:
+        class: consim\core\entity\News
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.news%
+
+    consim.core.entity.news_channel:
+        class: consim\core\entity\NewsChannel
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.news_channels%
+
+    consim.core.entity.news_topics:
+        class: consim\core\entity\NewsTopics
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.news_topics%
+
+    consim.core.entity.consim_user:
+        class: consim\core\entity\ConsimUser
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - @service_container
+            - %tables.consim.user%
+            - %tables.consim.figure%
+
+    consim.core.entity.consim_figure:
+        class: consim\core\entity\ConsimFigure
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.figure%
+
+    consim.core.entity.consim_skill:
+            class: consim\core\entity\Skill
+            scope: prototype # scope MUST be prototype for this to work!
+            arguments:
+                - @dbal.conn
+                - %tables.consim.skills%
+
+    consim.core.entity.consim_user_skill:
+            class: consim\core\entity\UserSkill
+            scope: prototype # scope MUST be prototype for this to work!
+            arguments:
+                - @dbal.conn
+                - %tables.consim.skills%
+                - %tables.consim.user_skills%
+
+    consim.core.entity.item:
+        class: consim\core\entity\Item
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.items%
+
+    consim.core.entity.inventory_item:
+            class: consim\core\entity\InventoryItem
+            scope: prototype # scope MUST be prototype for this to work!
+            arguments:
+                - @dbal.conn
+                - %tables.consim.inventory_items%
+                - %tables.consim.items%
+
+    consim.core.entity.anniversary:
+        class: consim\core\entity\Anniversary
+        scope: prototype # scope MUST be prototype for this to work!
+        arguments:
+            - @dbal.conn
+            - %tables.consim.anniversary%
