+++ core/controller/RequestController.php
@@ -1,73 +1,44 @@
 <?php
 /**
-*
-* @package ConSim for phpBB3.1
-* @copyright (c) 2015 Marco Candian (tacitus@strategie-zone.de)
-* @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
-*
-*/
+ *
+ * @package ConSim for phpBB3.1
+ * @copyright (c) 2015 Marco Candian (tacitus@strategie-zone.de)
+ * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
+ *
+ */
 
 namespace consim\core\controller;
 
+use consim\core\entity\Action;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
-* Main controller
-*/
-class Action
+ * Main controller
+ */
+class RequestController
 {
-	/** @var \phpbb\config\config */
-	protected $config;
-
-	/** @var ContainerInterface */
-	protected $container;
-
-	/** @var \phpbb\controller\helper */
-	protected $helper;
-
-	/** @var \phpbb\user */
-	protected $user;
-
-	/** @var \phpbb\template\template */
-	protected $template;
-
-	/** @var \phpbb\request\request */
-	protected $request;
-
-	/** @var \phpbb\db\driver\driver_interface */
-	protected $db;
-
 	/**
-	* Constructor
-	*
-	* @param \phpbb\config\config				$config				Config object
-	* @param \phpbb\controller\helper			$helper				Controller helper object
-	* @param ContainerInterface					$container			Service container interface
-	* @param \phpbb\user						$user				User object
-	* @param \phpbb\template\template			$template			Template object
-	* @param \phpbb\request\request				$request			Request object
-	* @param \phpbb\db\driver\driver_interface	$db					Database object
-	* @return \consim\core\controller\Action
-	* @access public
-	*/
-	public function __construct(\phpbb\config\config $config,
-								ContainerInterface $container,
-								\phpbb\controller\helper $helper,
-								\phpbb\user $user,
-								\phpbb\template\template $template,
-								\phpbb\request\request $request,
-								\phpbb\db\driver\driver_interface $db)
+	 * Constructor
+	 *
+	 * @param \phpbb\controller\helper			$helper				Controller helper object
+	 * @param ContainerInterface				$container			Service container interface
+	 * @param \phpbb\user						$user				User object
+	 * @param \phpbb\request\request			$request			Request object
+	 * @return \consim\core\controller\RequestController
+	 * @access public
+	 */
+	public function __construct(ContainerInterface $container,
+		\phpbb\controller\helper $helper,
+		\phpbb\user $user,
+		\phpbb\request\request $request)
 	{
-		$this->config = $config;
 		$this->container = $container;
 		$this->helper = $helper;
 		$this->user = $user;
-		$this->template = $template;
 		$this->request = $request;
-		$this->db = $db;
 	}
 
-	public function travel($travel_id)
+	public function startTravel($travel_id)
 	{
 		//Check the request
 		if (!$this->is_valid($travel_id) || !check_link_hash($this->request->variable('hash', ''), 'travel_' . $travel_id))
@@ -114,7 +85,6 @@ class Action
 		{
 			throw new \phpbb\exception\http_exception(403, 'NO_AUTH_OPERATION');
 		}
-
 		//Load ConsimUser
 		$consim_user = $this->container->get('consim.core.entity.consim_user')->load($this->user->data['user_id']);
 
@@ -181,7 +151,7 @@ class Action
 		}
 		$action->userDone();
 
-		redirect($this->helper->route('consim_core_action', array('action_id' => $action->getId())));
+		redirect($this->helper->route('consim_core_work', array('work_id' => $action->getId())));
 	}
 
 	protected function is_valid($value)
